serde: move is_serde_compatible to separate header
